Refactorings in preparation for unification of query and response.
1. newscatcher + newscatcher_cline => newscatcher + newscatcher::models
2. gnews + gnews_client => gnews + gnews::models
3. move some public top level parts to from client to lib (they wont change,
   client will be gone in follow up PR)
4. rename some functions using gnews to have gnews in the name
5. move code specific to newscatcher/gnews from client to the
   newscatcher/gnews modules
