+++ discovery_engine_core/core/src/engine.rs
@@ -533,7 +533,7 @@ where
                         excluded_sources: &excluded_sources,
                         filter,
                     };
-                    self.client.query_articles(&news_query).await
+                    self.client.query_gnews_articles(&news_query).await
                 }
                 SearchBy::Topic(topic) => {
                     let common = CommonQueryParts {
@@ -1102,7 +1102,7 @@ mod tests {
             .set_body_string(include_str!("../test-fixtures/newscatcher/duplicates.json"));
 
         Mock::given(method("GET"))
-            .and(path("/_lh"))
+            .and(path("/latest-headlines"))
             .respond_with(tmpl)
             .mount(&mock_server)
             .await;
